requirements.txt
anyio==4.9.0
APScheduler==3.11.0
blinker==1.9.0
certifi==2025.8.3
click==8.2.1
colorama==0.4.6
exceptiongroup==1.3.0
Flask==3.1.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.6
pandas==2.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==22.3
pytz==2025.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
Werkzeug==3.1.3
